Desenvolvimento da pasta service com base nos novos dados

+++ src/main/java/com/example/erpserver/DTOs/ItemProdutoDTO.java
@@ -19,5 +19,5 @@ public class ItemProdutoDTO {
     private UUID produtoId;
 
     @Min(value = 1, message = "Quantidade deve ser pelo menos 1")
-    private int quantidade;
+    private long quantidade;
 }
